change progress to 0-1

+++ lib/core/services/local/local_database/i_offline_database_provider.dart
@@ -5,7 +5,7 @@ import '../../../models/api/response.dart';
 import 'i_database_provider.dart';
 
 abstract class IOfflineDatabaseProvider extends IDatabaseProvider {
-  int progress;
+  double progress;
   Stream<Response> request(Request request);
   Future<bool> syncOnline(BuildContext context);
 }
